fix: allow sync.sh only on darwin

+++ scripts/sync.sh
@@ -3,9 +3,14 @@
 
 set -euo pipefail
 
-echo "Syncing repository to container"
-rsync -hrPti --delete --update --fsync --exclude buildroot /mnt/ /work 
+if [[ "${HOST_OS^^}" == "DARWIN" ]]; then
+    echo "Syncing repository to container"
+    rsync -hrPti --delete --update --fsync --exclude buildroot /mnt/ /work 
 
-# Sync container back to repository
-echo "Syncing container back to repository"
-rsync -hrPti --delete --update --fsync --exclude buildroot /work/ /mnt
+    # Sync container back to repository
+    echo "Syncing container back to repository"
+    rsync -hrPti --delete --update --fsync --exclude buildroot /work/ /mnt
+else
+    echo "sync.sh is only used on macOS"
+    exit 1
+fi
